Refactor deletion logic for Anime lists
This update refactors the 'delete' method in the Anime list controller and service. Previously, there was a hard-coded limit that prevented deletion of default anime lists. Now, this limitation has been moved to the service layer and checks whether an anime list is a default list before attempting deletion.

+++ backend/services/animeList.service.js
@@ -23,9 +23,17 @@ const animeListService = {
         return new AnimeListDto(animeListCreated);
     },
 
-    delete: async (animeLisId) => {
+    delete: async (animeListId) => {
+        const animeListFound = await db.AnimeList.findOne({
+            where: {id: animeListId},
+        })
+
+        if (animeListFound.isDefault) {
+            throw new Error('Cannot delete default anime list');
+        }
+
         const nbRowDeleted = await db.AnimeList.destroy({
-            where: {id: animeLisId}
+            where: {id: animeListId}
         });
 
         return nbRowDeleted;
